Add put request method in `HTTPClient` and `URLSessionHTTPClient`.
